requirements.txt
ctransformers==0.2.27
huggingface-hub==0.29.3
sentence-transformers==3.4.1
transformers==4.39.3
pinecone-client==2.2.4  
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-pinecone==0.2.3
langchain-huggingface==0.1.2
streamlit>=1.34.0
pypdf==4.1.0
python-dotenv==1.0.1
torch>=2.2.0 

